Remove falling damage logic dependency on PlayerMoveEvent
